+++ app/serializers/configured-citation-addon.ts
@@ -1,16 +1,6 @@
-import DS from 'ember-data';
-import { SingleResourceDocument } from 'osf-api';
+import ConfiguredAddonSerializer from './configured-addon';
 
-import GravyValetSerializer from './gravy-valet-serializer';
-
-export default class ConfiguredCitationAddonSerializer extends GravyValetSerializer {
-    serialize(snapshot: DS.Snapshot, options: {}) {
-        const serialized = super.serialize(snapshot, options) as SingleResourceDocument;
-        if (!serialized.data.attributes!.authorized_resource_uri) {
-            delete serialized.data.attributes!.authorized_resource_uri;
-        }
-        return serialized;
-    }
+export default class ConfiguredCitationAddonSerializer extends ConfiguredAddonSerializer {
 }
 
 declare module 'ember-data/types/registries/serializer' {
